Separate out refinement functions
